requirements.txt
Django==4.1.5
djangorestframework==3.14.0
environs==9.5.0
dj-database-url==1.2.0
django-filter==22.1